Data.Generically

Description

This module provides some generic functions to create instances of FromJSON, toJSON, and ToHttpApiData using DerivingVia. You can find usages in Web.Pixiv.Types.

HttpApiData

newtype CustomHttpApiData t a

A wrapper to derive ToHttpApiData instances with modifications.

Constructors

CustomHttpApiData

Fields

unCustomHttpApiData :: a

class CustomHttpApiDataOption xs where

Reifies StringModifiers applied on CustomHttpApiData.

Methods

option :: String -> String

class GToHttpApiData f where

Class of generic representation types that can be converted to query param.

Methods

gToQueryParam :: f p -> (String -> String) -> Text

type PixivHttpApiData (k :: Symbol) = CustomHttpApiData '[StripPrefix k, CamelToSnake]

Strips prefix k from query parameter's name then converts it to snake case.

type PixivHttpApiData' = PixivHttpApiData ""

Converts query parameter's name to snake case.

JSON

type PixivJSON (k :: Symbol) = CustomJSON '[FieldLabelModifier (PixivLabelModifier k, CamelToSnake)]

Strips prefix _ and k from fields' labels then converts it to snake case, making sure results are non-empty.

type EnumJSON (k :: Symbol) = CustomJSON '[ConstructorTagModifier (StripPrefix k, CamelToSnake)]

Strips prefix k from constructors' tags then converts it to snake case.

type PixivJSON' = PixivJSON ""

Strips prefix _ from fields' labels then converts it to snake case.

type EnumJSON' = EnumJSON ""

Converts constructors' tags to snake case.

data PixivLabelModifier (k :: Symbol)

A StringModifier that strips prefix _ and k, and returns non-empty String.

Re-export

class ToHttpApiData a where

Minimal complete definition

toUrlPiece | toQueryParam

Methods

toUrlPiece :: a -> Text

toEncodedUrlPiece :: a -> Builder

toHeader :: a -> ByteString

toQueryParam :: a -> Text