Web.Pixiv.API

Description

Functions to access pixiv api. They are supposed to be run in Web.Pixiv.Types.PixivT.

You may notice that except getUserBookmarks, many of functions take a page as input. This is because each query contains 30 entries, i.e. you should pass 2 if you want items ranged from 31 to 60.

Trending

getTrendingTags

Arguments

:: MonadPixiv m

=> Maybe Bool
includes translated tag results

-> m [TrendingTag]

Gets trending tags.

getRecommendedIllusts

Arguments

:: MonadPixiv m

=> Maybe Bool
includes privacy policy

-> Maybe Bool
includes translated tag results

-> m [Illust]

Gets recommended illustrations of the account which is used for login.

getRecommendedMangas :: MonadPixiv m => Maybe Bool -> Maybe Bool -> m [Illust]

Gets recommended mangas of the account which is used for login.

Illust

getIllustDetail

Arguments

:: MonadPixiv m

=> Int
illust id

-> m Illust

Gets the details of a illustration.

getIllustComments

Arguments

:: MonadPixiv m

=> Int
illust id

-> Int
page

-> m Comments

Gets the comments of a illustration.

getIllustRelated

Arguments

:: MonadPixiv m

=> Int
illust id

-> Int
page

-> m [Illust]

Gets related illustrations.

getIllustRanking

Arguments

:: MonadPixiv m

=> Maybe RankMode
rank mode

-> Int
page

-> m [Illust]

Gets ranking illustrations.

getIllustFollow

Arguments

:: MonadPixiv m

=> Publicity
restrict

-> Int
page

-> m [Illust]

Gets illustrations of artists which the login account follows.

getIllustNew

Arguments

:: MonadPixiv m

=> Int
page

-> m [Illust]

Gets the newest illustrations.

getUgoiraMetadata

Arguments

:: MonadPixiv m

=> Int
illust id

-> m UgoiraMetadata

Gets the metadata of a ugoira

Search

searchIllust

Arguments

:: MonadPixiv m

=> SearchTarget
search target

-> Text
word

-> Maybe Bool
includes translated tag

-> Maybe SortingMethod
sorting method

-> Maybe Duration
duration

-> Int
page

-> m [Illust]

Searches an illustration.

searchUser

Arguments

:: MonadPixiv m

=> Text
word

-> Maybe SortingMethod
sorting method

-> Int
page

-> m [UserPreview]

Searches an user.

User

getUserDetail

Arguments

:: MonadPixiv m

=> Int
user id

-> m UserDetail

Gets the details of an user.

getUserIllusts

Arguments

:: MonadPixiv m

=> Int
user id

-> Maybe IllustType
illust type

-> Int
page

-> m [Illust]

Gets illustrations submitted by a specific user.

getUserFollowing

Arguments

:: MonadPixiv m

=> Int
user id

-> Publicity
restrict

-> Int
page

-> m [UserPreview]

Gets the following of an user.

getUserFollower

Arguments

:: MonadPixiv m

=> Int
user id

-> Int
page

-> m [UserPreview]

Gets the followers of an user.

getUserMypixiv

Arguments

:: MonadPixiv m

=> Int
user id

-> Int
page

-> m [UserPreview]

Gets mypixiv of an user.

getUserBookmarks

Arguments

:: MonadPixiv m

=> Int
user id

-> Publicity
restrict

-> Maybe Int
the last illust id of this query

-> m ([Illust], Maybe Int)

Gets illustrations collected by a specific user. Returns the a list of illustrations and result and an integer for paging.