Web.Pixiv.Auth

Description

Authentication pixiv API. Users should not use logics in this module directly, since Web.Pixiv.Types.PixivT takes over token management, providing user friendly operations.

Documentation

newtype Token

A wrapped Text represents a token.

Constructors

Token

Fields

unToken :: Text

data Credential

Authentication credentials for pixiv API.

Normally, users are supposed to create value of this data type using Password constructor, then pass it to runPixivT.

Constructors

Password

Fields

username :: ByteString

password :: ByteString

RefreshToken

Fields

cr_refreshToken :: Token

data OAuth2Token

Successful result.

Constructors

OAuth2Token

Fields

oa_accessToken :: Token

oa_expiresIn :: Int

oa_refreshToken :: Token

data OAuth2Error

Failed result.

Constructors

OAuth2Error

Fields

oa_error :: Errors

oa_message :: Text

data OAuth2Result

Authentication result.

Constructors

AuthSuccess OAuth2Token

AuthFailure OAuth2Error

data Errors

Authentication failure reasions.

Constructors

InvalidRequest

InvalidClient

InvalidGrant

UnauthorizedClient

UnsupportedGrantType

InvalidScope

auth :: Manager -> Credential -> IO OAuth2Result

Given a credential, performs a authentication request.

auth' :: Manager -> Credential -> IO OAuth2Token

Like auth, but immediately throws OAuth2Error if auth failed.