Web.Pixiv.Download

Description

A set of utilities for downloading pixiv things.

DownloadM monad

type DownloadM = MaybeT (ReaderT Manager IO)

DownloadM monad is a synonym for IO computation wrapped in Manager environment, which may exit without producing value, indicating the download failed.

liftMaybe :: Maybe a -> DownloadM a

Lifts a pure Maybe value to DownloadM.

liftToPixivT :: MonadIO m => DownloadM a -> PixivT m (Maybe a)

Lifts a download computation to PixivT.

DownloadM needs Manager to perform the download, which can be provided by TokenState.

runDownloadM :: Manager -> DownloadM a -> IO (Maybe a)

Executes the download computation in IO.

Download actions

downloadPixiv :: Text -> DownloadM ByteString

Downloads something in DownloadM, given url.

downloadSingleIllust :: Illust -> DownloadM ByteString

Downloads a single page illust.

Chooses the first one if the illust has many pages, preferring high quality images. Returns Nothing if can't find any image url.

downloadUgoiraToMP4

Arguments

:: UgoiraMetadata
Information of ugoira to download

-> Maybe FilePath
Path to ffmpeg

-> DownloadM (String, ByteString)

Downloads UgoiraFrames, then converts it to MP4 calling external ffmpeg.