Web.Pixiv.Types

Description

Data types used in API. This module enables DuplicateRecordFields, so please consider using Web.Pixiv.Types.Lens to access fields smoothly.

ImageUrl

type ImageUrl = Text

Image urls are represented in Text.

data ImageUrls

An object contains image urls.

Example in Illust or MetaPage: "image_urls": { "square_medium": "...", "medium": "...", "large": "..." }

Example in User: "profile_image_urls": { "medium": "..." }

Constructors

ImageUrls

Fields

_squareMedium :: Maybe ImageUrl

_medium :: Maybe ImageUrl

_large :: Maybe ImageUrl

_original :: Maybe ImageUrl

newtype OriginalImageUrl

An object contains a single image url.

In Illust: "meta_single_page": { "original_image_url": "..." }

Constructors

OriginalImageUrl

Fields

_originalImageUrl :: Maybe ImageUrl

Tag

data Tag

A tag.

Example: { "name": "...", "translated_name": null }

Constructors

Tag

Fields

_name :: Text

_translatedName :: Maybe Text

data TrendingTag

A trending tag.

Don't confuse with Tag. TrendingTag contains Illust, and the textual name of the tag is called tag, instead of name in Tag.

Constructors

TrendingTag

Fields

_trendTag :: Text

_translatedName :: Maybe Text

_illust :: Illust

newtype TrendingTags

A wrapper of TrendingTags for JSON deserialization.

Constructors

TrendingTags

Fields

_trend_tags :: [TrendingTag]

Illustration

data Series

Manga series.

Constructors

Series

Fields

_seriesId :: Int

_title :: Text

data IllustType

Type of illustration.

In pixiv API, all of illustrations, mangas, and ugoiras are represented in Illust data type. So they can be distinguished by IllustType.

Constructors

TypeIllust

TypeManga

TypeUgoira

newtype MetaPage

A page of Illust containing ImageUrls.

Constructors

MetaPage

Fields

_imageUrls :: ImageUrls

data Illust

Illustraion data type.

See IllustType.

Constructors

Illust

Fields

_illustId :: Int

_title :: Text

_illustType :: IllustType

_imageUrls :: ImageUrls

_caption :: Text

_restrict :: Int

_user :: User

_tags :: [Tag]

_tools :: [Text]

_createDate :: UTCTime

_pageCount :: Int

_width :: Int

_height :: Int

_sanityLevel :: Int

_xRestrict :: Int

_series :: Maybe Series

Only manga may have this field.

_metaSinglePage :: Maybe OriginalImageUrl

Only single page illustration has this field.

_metaPages :: [MetaPage]

Only multi page illustration has this field.

_totalView :: Int

_totalBookmarks :: Int

_isBookmarked :: Bool

For login account.

_visible :: Bool

_isMuted :: Bool

_totalComments :: Maybe Int

data Illusts

Response of API which returns illustrations.

Constructors

Illusts

Fields

_illusts :: [Illust]

_nextUrl :: Maybe Text

newtype IllustWrapper

A wrapper of Illust for JSON deserialization.

Constructors

IllustWrapper

Fields

_illust :: Illust

User

data User

User data type.

Constructors

User

Fields

_userId :: Int

_name :: Text

_account :: Text

_profileImageUrls :: ImageUrls

_comment :: Maybe Text

_isFollowed :: Maybe Bool

For login account.

data UserProfile

UserProfile data type.

Not sure if all fields are covered, and maybe some fields should be optional.

Constructors

UserProfile

Fields

_webpage :: Text

_gender :: Text

_birth :: Text

_birthDay :: Text

_birthYear :: Int

_region :: Text

_addressId :: Int

_countryCode :: Text

_job :: Text

_jobId :: Int

_totalFollowUsers :: Int

_totalMypixivUsers :: Int

_totalManga :: Int

_totalIllustBookmarksPublic :: Int

_totalIllustSeries :: Int

_totalNovelSeries :: Int

_backgroundImageUrl :: ImageUrl

_twitterAccount :: Text

_twitterUrl :: Text

_pawooUrl :: Maybe Text

_isPreminum :: Maybe Bool

_isUsingCustomProfileImage :: Bool

data Publicity

Publicity data type.

The value public or private are present in ProfilePublicity. This type is also used in restrict query param.

Constructors

Public

Private

data ProfilePublicity

Profile publicity of a user.

Not sure if all fields are covered, and maybe some fields should be optional.

Constructors

ProfilePublicity

Fields

_gender :: Publicity

_region :: Publicity

_birthDay :: Publicity

_birthYear :: Publicity

_job :: Publicity

_pawoo :: Bool

data Workspace

Workspace information of a user. Not sure if all fields are covered, and maybe some fields should be optional.

Constructors

Workspace

Fields

_pc :: Text

_monitor :: Text

_tool :: Text

_scanner :: Text

_tablet :: Text

_mouse :: Text

_printer :: Text

_desktop :: Text

_music :: Text

_desk :: Text

_chair :: Text

_comment :: Text

_workspaceImageUrl :: Maybe Text

data UserDetail

Details of a user.

Constructors

UserDetail

Fields

_user :: User

_profile :: UserProfile

_profilePublicity :: ProfilePublicity

_workspace :: Workspace

data UserPreview

A preview of user information.

Except getUserDetail, other API involving users return this data type.

Constructors

UserPreview

Fields

_user :: User

_illusts :: [Illust]

_novels :: Value

Novels are not supported currently.

_isMuted :: Bool

data UserPreviews

Response of API which returns user previews.

Constructors

UserPreviews

Fields

_userPreviews :: [UserPreview]

_nextUrl :: Maybe Text

Comment

data Comment

A comment.

Constructors

Comment

Fields

_commentId :: Int

_comment :: Text

_date :: UTCTime

_user :: User

_parentComment :: Maybe Value

data Comments

Response of API which returns comments.

Constructors

Comments

Fields

_totalComments :: Int

_comments :: [Comment]

_nextUrl :: Maybe Text

NextUrl

type family NextUrlLess a

Undecorate next_url of a type.

next_url is returned by some APIs for paging.

class HasNextUrl a where

Class to get or unwrap next_url.

Methods

unNextUrl :: a -> NextUrlLess a

getNextUrl :: a -> Maybe Text

Ugoria

data UgoiraFrame

A ugoira frame.

Constructors

UgoiraFrame

Fields

_ugoiraFile :: Text

File name.

_ugoiraDelay :: Int

Duration of this frame (in millisecond).

newtype ZipUrls

A wrapper of ugoira zip file url.

Constructors

ZipUrls

Fields

_zipMedium :: Text

data UgoiraMetadata

Ugoira is a frame animation, whose common information is represented in Illust. This metadata contains a link to download the zip archive, which compresses frames of the ugoira; and UgoiraFrames to represents metadata of each frame. Using getUgoiraMetadata can obtain value of this type. See ugoiraMetadataToFFConcat and downloadUgoiraToMP4.

Constructors

UgoiraMetadata

Fields

_zipUrls :: ZipUrls

_frames :: [UgoiraFrame]

newtype UgoiraMetadataWrapper

A wrapper of UgoiraMetadata for JSON deserialization.

Constructors

UgoiraMetadataWrapper

Fields

_ugoiraMetadata :: UgoiraMetadata

Article

data SpotlightArticle

Spotlight article.

Constructors

SpotlightArticle

Fields

_saId :: Int

_title :: Text

_pureTitle :: Text

_thumbnail :: Text

_articleUrl :: Text

_publishDate :: UTCTime

_category :: Text

_subcategoryLabel :: Text

data SpotlightArticles

Response of API which returns spotlight articles.

Constructors

SpotlightArticles

Fields

_spotlightArticles :: [SpotlightArticle]

_nextUrl :: Maybe Text

Http

data RankMode

Rank mode query parm.

See getIllustRanking

Constructors

Day

DayR18

DayMale

DayMaleR18

DayFemale

DayFemaleR18

Week

WeekR18

WeekR18G

Month

WeekOriginal

WeekRookie

DayManga

data SearchTarget

Search target query parm.

See searchIllust.

Constructors

ExactMatchForTags

PartialMatchForTags

TitleAndCaption

data SortingMethod

Sorting method query parm.

See searchIllust and searchUser.

Constructors

DateDesc

DateAsc

data Duration

Duration query parm.

See searchIllust.

Constructors

WithinLastDay

WithinLastMonth

WithinLastYear