Web.Pixiv.Types.PixivT

Description

The core monad of this library. PixivT maintains a pixiv login state, and provides an environment to perform computations created by servant.

ClientT monad transformer

newtype ClientT m a

Transformer version of ClientM, changing the base IO to m.

Constructors

ClientT

Fields

unClientT :: ReaderT ClientEnv (ExceptT ClientError m) a

runClientT :: ClientEnv -> ClientT m a -> m (Either ClientError a)

Executes a computation in the client monad.

MonadPixiv class

class (RunClient m, MonadIO m) => MonadPixiv m where

The mtl-style class of pixiv monad.

Methods

takePixivState :: m PixivState

Reads the stored PixivState, when used in a multithreaded setting, this should block all other thread from reading the PixivState until putPixivState is called.

putPixivState :: PixivState -> m ()

Writes a new PixivState.

readPixivState :: m PixivState

Reads the stored PixivState, without blocking other threads which want to read this state.

Don't confuse with takePixivState, please refer to readMVar.

data PixivState

State stored in MonadPixiv.

Constructors

PixivState

Fields

tokenState :: TokenState

acceptLanguage :: Maybe Text

PixivT monad transformer

newtype PixivT m a

A thread safe implementation of MonadPixiv.

Constructors

PixivT

Fields

unPixivT :: ReaderT (MVar PixivState) (ClientT m) a

liftC :: Monad m => ClientT m a -> PixivT m a

Lifts a computation in ClientT to PixivT.

runPixivT :: MonadIO m => Manager -> Credential -> PixivT m a -> m (Either ClientError a)

Interprets the PixivT effect, with a supplied Manager.

runPixivT' :: MonadIO m => Credential -> PixivT m a -> m (Either ClientError a)

Like runPixivT, but creates a new Manager everytime.

Token

data TokenState

Pixiv auth state.

Constructors

TokenState

Fields

accessToken :: Token

Token to access pixiv api.

refreshToken :: Token

Token to obtain new accessToken without giving username and password.

expirationTime :: UTCTime

Time stamp when both accessToken and refreshToken are invalid.

manager :: Manager

computeTokenState

Arguments

:: Manager

-> Credential
Could be username with password or refreshToken.

-> UTCTime
Current time.

-> IO TokenState

Computes the TokenState.

Utilities

getAccessToken :: MonadPixiv m => m Token

Retrieves the accessToken from pixiv monad.

If the token is overdue, it will call computeTokenState to refresh.

getAccessTokenWithAccpetLanguage :: MonadPixiv m => m (Token, Maybe Text)

Retrieves the accessToken and acceptLanguage in one go.